requirements.txt
python-docx
requests
PyPDF2